Struct cam::Camera

pub struct Camera<T = f32> {
    pub position: Vector3<T>,
    pub up: Vector3<T>,
    pub right: Vector3<T>,
    pub forward: Vector3<T>,
}

Models a camera with position and directions.

Fields

position

The camera position.

up

The up direction.

right

The right direction.

forward

The forward direction.

Methods

impl<T: Float + FloatMath + Copy> Camera<T>

fn new(x: T, y: T, z: T) -> Camera<T>

Constructs a new camera.

Places the camera at [x, y, z], looking towards pozitive z.

fn orthogonal(&self) -> Matrix4<T>

Computes an orthogonal matrix for the camera.

This matrix can be used to transform coordinates to the screen.

fn look_at(&mut self, x: T, y: T, z: T)

Orients the camera to look at a point.

fn set_yaw_pitch(&mut self, yaw: T, pitch: T)

Sets yaw and pitch angle of camera in radians.